rename population files to place-population
