+++ mcp-server/src/types/place-population.ts
